move lxml dependent methods to separate file
